requirements.txt
py-trello
crowdin-api-client